+++ bot/match/processes/__init__.py
@@ -4,6 +4,4 @@ from .faction_picking import FactionPicking
 from .getting_ready import GettingReady
 from .match_playing import MatchPlaying
 from .player_picking import PlayerPicking
-from .getting_ready_2 import GettingReady2
-__all__ = ["BasePicking", "CaptainSelection", "FactionPicking", "GettingReady", "MatchPlaying", "PlayerPicking",
-           "GettingReady2"]
+__all__ = ["BasePicking", "CaptainSelection", "FactionPicking", "GettingReady", "MatchPlaying", "PlayerPicking"]